renamed files better and removed components
